backend/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
Flask-Cors==4.0.0
reportlab==4.0.8
pytest==8.0.0
pytest-cov==4.1.0

